changed activity to wearableActivity, added AmbinetMode settings , fixed weatherId data not propagating correctly

+++ app/src/main/java/com/example/android/sunshine/app/ForecastAdapter.java
@@ -159,7 +159,7 @@ public class ForecastAdapter extends RecyclerView.Adapter<ForecastAdapter.Foreca
     public void sendWeatherData(int weatherId, double high, double low, String friendlyDayString) {
 
         PutDataMapRequest dataMap = PutDataMapRequest.create("/weather-data");
-        dataMap.getDataMap().putLong("WEATHER_ID_KEY", weatherId);
+        dataMap.getDataMap().putInt("WEATHER_ID_KEY", weatherId);
         dataMap.getDataMap().putInt("WEATHER_HIGH_KEY", new Double(high).intValue());
         dataMap.getDataMap().putInt("WEATHER_LOW_KEY",  new Double(low).intValue());
         dataMap.getDataMap().putString("WEATHER_CURRENT_DATE",friendlyDayString);
